[fix](vllm): address upgrade review feedback
Keep vLLM-only behavior out of native models and centralize the pinned vLLM release metadata used by CI.
