+++ .github/vllm-version.env
@@ -1,0 +1,2 @@
+VLLM_COMMIT=752a3a504485790a2e8491cacbb35c137339ad34
+VLLM_VERSION=0.25.1
