+++ .github/workflows/atom-vllm-benchmark.yaml
@@ -341,18 +341,24 @@ jobs:
           BENCHMARK_CLIENT="${INPUT_BENCHMARK_CLIENT:-InferenceMax bench}"
 
           mapfile -t VLLM_MET < <(python3 - <<'PY'
-          import re
           from pathlib import Path
 
-          text = Path(".github/workflows/docker-release.yaml").read_text(encoding="utf-8")
-          commit_match = re.search(r'^\s*VLLM_COMMIT:\s*"([^"]+)"', text, re.MULTILINE)
-          version_match = re.search(r'^\s*VLLM_VERSION:\s*"([^"]+)"', text, re.MULTILINE)
+          config = {}
+          for line in Path(".github/vllm-version.env").read_text(
+              encoding="utf-8"
+          ).splitlines():
+              key, separator, value = line.partition("=")
+              if separator and key:
+                  config[key] = value
 
-          if not commit_match or not version_match:
-              raise SystemExit("Failed to read VLLM_COMMIT/VLLM_VERSION from .github/workflows/docker-release.yaml")
+          try:
+              print(config["VLLM_COMMIT"])
+              print(config["VLLM_VERSION"])
+          except KeyError as error:
+              raise SystemExit(
+                  f"Missing {error.args[0]} in .github/vllm-version.env"
+              ) from error
 
-          print(commit_match.group(1))
-          print(version_match.group(1))
           PY
           )
 
@@ -428,7 +434,7 @@ jobs:
             printf -- '\nBecause an explicit `oot_image` input was provided, this run uses that image directly and skips branch-based rebuild logic.\n' \
               >> "$GITHUB_STEP_SUMMARY"
           elif [[ "${REBUILD_OOT_IMAGE}" == "true" ]]; then
-            printf -- '- Base image: `%s`\n- vLLM version: `%s`\n- vLLM commit: `%s`\n- vLLM source: `.github/workflows/docker-release.yaml`\n' \
+            printf -- '- Base image: `%s`\n- vLLM version: `%s`\n- vLLM commit: `%s`\n- vLLM source: `.github/vllm-version.env`\n' \
               "${OOT_BASE_IMAGE}" \
               "${SELECTED_VLLM_VERSION}" \
               "${SELECTED_VLLM_COMMIT}" >> "$GITHUB_STEP_SUMMARY"
