+++ .github/workflows/atom-vllm-test.yaml
@@ -30,7 +30,6 @@ env:
   GITHUB_REPO_URL: ${{ github.event.pull_request.head.repo.clone_url || 'https://github.com/ROCm/ATOM.git' }}
   GITHUB_COMMIT_SHA: ${{ github.event.pull_request.head.sha || github.event.head_commit.id || github.sha }}
   NIGHTLY_OOT_IMAGE_TAG: rocm/atom-dev:vllm-latest
-  VLLM_COMMIT: "752a3a504485790a2e8491cacbb35c137339ad34"
 
 jobs:
   # ---------- OOT CI path (pull_request only) ----------
@@ -176,8 +175,6 @@ jobs:
     timeout-minutes: 180
     env:
       CONTAINER_NAME: atom_vllm_oot_${{ strategy.job-index }}
-      VLLM_COMMIT: 752a3a504485790a2e8491cacbb35c137339ad34
-      VLLM_VERSION: "0.25.1"
       AITER_ARTIFACT_ID: ${{ needs.download_aiter_wheel.outputs.aiter_artifact_id }}
 
     steps:
@@ -208,6 +205,9 @@ jobs:
 
       - name: Checkout ATOM repo
         uses: actions/checkout@v4
+
+      - name: Load vLLM version
+        run: cat .github/vllm-version.env >> "$GITHUB_ENV"
 
       - name: Detect ATOM base image changes
         id: atom-base-changes
