+++ .github/workflows/docker-release.yaml
@@ -95,8 +95,6 @@ jobs:
       RCCL_REPO: "https://github.com/ROCm/rccl.git"
       RCCL_BRANCH: "29e1567b95e28823b0beb1a988adc587bfab5b4f"
       GPU_ARCH: "gfx942;gfx950"
-      VLLM_COMMIT: "752a3a504485790a2e8491cacbb35c137339ad34"
-      VLLM_VERSION: "0.25.1"
       SGLANG_REPO: "https://github.com/sgl-project/sglang.git"
       SGLANG_REF: "v0.5.15.post1"
       SGLANG_VERSION: "0.5.15.post1"
@@ -107,6 +105,9 @@ jobs:
 
       - name: Checkout ATOM repo
         uses: actions/checkout@v6
+
+      - name: Load vLLM version
+        run: cat .github/vllm-version.env >> "$GITHUB_ENV"
 
       - name: Login to Docker Hub
         uses: ./.github/actions/docker-auth
@@ -266,7 +267,7 @@ jobs:
           docker build --network=host -t atom_oot_release:ci \
             --build-arg OOT_BASE_IMAGE="${OOT_BASE_IMAGE}" \
             --build-arg MAX_JOBS=64 \
-            --build-arg VLLM_COMMIT="${{ env.VLLM_COMMIT }}" \
+            --build-arg VLLM_COMMIT="${VLLM_COMMIT}" \
             --build-arg INSTALL_LM_EVAL=1 \
             --build-arg INSTALL_FASTSAFETENSORS=1 \
             -f docker/vllm_release.dockerfile .
@@ -285,7 +286,7 @@ jobs:
         if: ${{ success() && steps.gates.outputs.oot == 'true' }}
         run: |
           set -euo pipefail
-          VLLM_VER="${{ env.VLLM_VERSION }}"
+          VLLM_VER="${VLLM_VERSION}"
           OOT_TAG_SUFFIX="${{ inputs.oot_tag_suffix || '' }}"
           OOT_TAG="vllm-v${VLLM_VER}-nightly_$(date +%Y%m%d)"
           OOT_LATEST_TAG="vllm-latest"
